tests/requirements.txt
pytest>=7.0.0
requests>=2.28.0
kafka-python>=2.0.0
pymongo>=4.0.0
dvc>=2.0.0
wandb>=0.13.0
prefect>=2.0.0
torch>=2.0.0
numpy>=1.22.0
scikit-learn>=1.0.0
fastapi>=0.95.0
termcolor>=2.0.0
pytest-cov>=4.0.0
matplotlib>=3.5.0